Avoid to find a record in the DB during spec execution
This controller spec was randomly failing in Travis.
So we decided to remove explicit calls to AR objects during the test
execution. The api endpoint returns more data so that all the existing expectations
can be verified.

The modification of the data outputed by the api endpoint was reflected in the JS
code that consumes this endpoint.