Remove zeus and guard-zeus from the bundle
Having guard-zeus specified as a dependency is problematic because
it also pulls in the zeus gem as a dependency. This makes it impo-
ssible to update or switch zeus versions locally without affecting
other developers.

The maintainers of zeus say that it is designed to be installed and
run outside of the bundle: https://github.com/burke/zeus