Add Skylight instrumentation
If you’re not already familiar with [Skylight](https://www.skylight.io),
it is a smart profiler for Rails apps. Skylight makes it easy to
pinpoint performance issues in Rails applications.

We work on a lot of open source projects ourselves, and in our
experience it can be pretty hard to get contributors to work on
application performance issues. Few contributors consider working
on performance problems, and the ones that might be interested may
not even know where to start.

By making performance information more accessible, we hope to
inspire potential contributors to tackle slow parts of your app,
and have a good way to see if their contributions helped.

Local deployments that are interested in this can request an API
token from https://www.skylight.io/oss and set it in
`config/application.yml` (if no token is set, it will log a message to
`log/skylight.log` but won't otherwise prevent the app from working
normally).