Upgrade Spree to 2.0.4 instead of 2.0.13
Although we don't know how big the fixes introduced between 2.0.4 and
2.0.13 are, jumping from roughly 1.3.99 to 2.0.13 seems too big of
a change to do in one go.

Unless more issues like this one come up that force us to upgrade to
a newer version I believe it's best to be conservative. You'll find an
in-depth explanation in
https://github.com/openfoodfoundation/openfoodnetwork/wiki/Spree-2.0-upgrade#getting-to-version-204-

A previous effort was done to upgrade to v2.0.13 in
https://github.com/openfoodfoundation/openfoodnetwork/pull/2209