Change sorting to be done in ascending order first
Currently, we always toggle "reverse" when triggering a sort.

If "reverse" is initially set to false, triggering a sort for the first
time then toggles this to true. The effect is that, the first time that
sorting is done, the rows are sorted in reverse order.

This is not intuitive - they should be sorted in ascending order first.