Replace feature spec with view spec
There is not need to boot a browser and hit all stack layers to check
how we display information in a template.

Furthermore, relying on feature specs by default not only leads to long running
and fragile test suites but also makes it impossible to have a proper
testing pyramid.

While the former feature spec took 0.52272 seconds to finish, the new
view spec took 0.39279 seconds. Nearly a 25% gain.