Move Variant deletion into its own service
This keeps the override of Spree's model leaner. More importantly, it
prepares us for using `destroy` instead of `delete`.

In the past, `Product#delete` soft-deleted the product, but didn't
delete the variants. When we use `Product#destroy` to soft-delete the
product, it will also call destroy on the variants. If the model doesn't
allow the deletion of the last variant, it will fail. So when a product
is deleted we want to allow the deletion of all variants. But the user
should not be allowed to delete the last variant. That's why I'm moving
the check to the controller level.

Related commits:

- https://github.com/openfoodfoundation/openfoodnetwork/commit/e6c7acdff356d02233d5f993cb93c2d1af824548
- https://github.com/openfoodfoundation/openfoodnetwork/commit/2b47c9145a9b18f346dd186ef4254dd062eb0648
- https://github.com/openfoodfoundation/openfoodnetwork/commit/b9f19d5777ed5034297dd3725c17bccffd39cf5d#diff-412c5af2ec1ba9f6643f6df5a673c1d4R105