Add rake task to empty a distributor enterprise
This wipes out all proxy orders, orders, inventory units and line items
of the specified distributor enterprise.

This is useful for the case of a hub that's been trying the platform and
then wants to clean up the mess to proceed to regular operation. This
also ensures we don't charge them for these initial tests.