Add rake tasks to remove OCs
Adds one to remove a single order cycle and another one that removes all
the order cycles of the specified coordinator.

This is useful for the case of a hub that's been trying the platform and then
wants to clean up the mess to proceed to regular operation. This also ensures
we don't charge them for these initial tests.