Use upstream version of simple_form and re-add deleted initializer
The current modified version we're using is incompatible with Rails 4.

The note says: "Fix bug in simple_form preventing collection_check_boxes usage within form_for block. When merged, revert to upstream gem"

I checked, and the commit that this old fork was designed for was actually merged later the same day (19 Jun 2012)!