Update new adjustment totals to work with fees
The `adjustment_total` column here is mainly relevant to OFN fees, where the adjustment (fee) might be $2.0 and the tax on it is $0.2. Spree doesn't use this, but it does have something similar called `promo_total` which includes calculations based on deductions from promotions (which we don't have in OFN).