Extract update logic needed after a shipping rate has been selected/changed
Spree seems to want the selected shipping rate to only be changed via `Shipment#update_attributes_and_order`, but we also change it via `Shipment#selected_shipping_rate_id=`. The extracted logic here needs to be run any time the selected shipping rate changes. If we don't do this, the shipment will have zero cost, and the order totals will not be updated.