Allow toggling a feature to all users
This comes in handy while in development and you don't want to check
which user logged in. Simply run the app with `BETA_TESTERS=all` and
solved.

It also give us the flexibility to toggle it for all users in production
without needing a deploy. Just change the ENV var and restart (reload
won't be enough).