Move invoice-related specs to their own file
The `Spree::Admin::OrdersController` is quite long but it'll grow more
in the upcoming commits. That's just a symptom of that controller having
too many responsibilities. It does much more than CRUD on `Spree::Order`
(the Rails convention).

Things like invoices are an entity on their own and would better fit
into a `InvoicesController`. Hopefully, splitting up the tests is hint
for the next dev to do that.